+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixed issue in HTMLAudioSound where `mute` would get into a recursive loop.
 * Every RenderTexture would draw the same content due to a mis-use of the CanvasPool (this also impacted TileSprites). Fix #3555 (thanks @kuoruan)
 * Group.add and Group.addMultiple now respect the Group.maxSize property, stopping you from over-populating a Group (thanks @samme)
+* When using HTML5 Audio, sound manager now tries to unlock audio after every scene loads, instead of only after first one. Fix #3309 (thanks @pavle-goloskokovic)
 
 ### Updates